schema-builder: minor fixes. extension types.
