+++ packages/schema-builder/lib/gql-schema-builder/parsePermissions.ts
@@ -3,8 +3,15 @@ import { extensions as defaultExtensions } from "./extensions";
 import { createSchemaBasics } from "./createSchemaBasics";
 import { createMutation } from "./createMutation";
 import { IPermissionContext, IPermission, IConfig, IResolverMeta } from "./interfaces";
+import { IParser } from "./extensions/interfaces";
+import { DefinitionNode, DocumentNode } from "graphql";
 
-export function parsePermissions(permissions: IPermission[], permissionContext: IPermissionContext, extensions, config: IConfig) {
+export function parsePermissions(
+  permissions: IPermission[],
+  permissionContext: IPermissionContext,
+  extensions: IParser[],
+  config: IConfig
+): { gqlDocument: DocumentNode; meta: IResolverMeta; sql: any[] } {
   const meta: IResolverMeta = {
     query: {},
     mutation: {},
@@ -12,11 +19,11 @@ export function parsePermissions(permissions: IPermission[], permissionContext: 
   };
 
   // TODO: Dustin: evaluate: permissionContext.gqlDocument = [...permissionContext.gqlDocument, ...createSchemaBasics()];
-  createSchemaBasics().forEach((d) => permissionContext.gqlDocument.definitions.push(d));
+  createSchemaBasics().forEach((d) => (permissionContext.gqlDocument.definitions as DefinitionNode[]).push(d));
 
   const sql = [];
   // TODO: Dustin: same story... evaluate: permissionContext.gqlDocument = [...permissionContext.gqlDocument, ...createSchemaBasics()];
-  const currentExtensions = extensions.slice().concat(defaultExtensions.slice());
+  const currentExtensions: IParser[] = extensions.slice().concat(defaultExtensions.slice());
 
   permissions.forEach((permission) => {
     const result = parsePermission(permission, permissionContext, currentExtensions, config);
@@ -35,7 +42,7 @@ export function parsePermissions(permissions: IPermission[], permissionContext: 
   Object.values(meta.mutation).forEach((mutation) => {
     const extendArguments = [];
     let myMutation = mutation;
-    currentExtensions.forEach((parser: any) => {
+    currentExtensions.forEach((parser) => {
       if (parser.modifyMutation != null) {
         const result = parser.modifyMutation(myMutation);
         if (result != null) {
@@ -51,19 +58,19 @@ export function parsePermissions(permissions: IPermission[], permissionContext: 
       }
     });
     const gqlMutation = createMutation(myMutation.name, myMutation.gqlReturnTypeName, myMutation.gqlInputTypeName, extendArguments);
-    permissionContext.gqlDocument.definitions.push(gqlMutation);
+    (permissionContext.gqlDocument.definitions as DefinitionNode[]).push(gqlMutation);
     modifiedMutation[myMutation.name] = mutation;
   });
 
   meta.mutation = modifiedMutation;
 
   // Loop over extensions to add definitions
-  currentExtensions.forEach((parser: any) => {
+  currentExtensions.forEach((parser) => {
     if (parser.extendDefinitions != null) {
       const definitions = parser.extendDefinitions(permissionContext.gqlDocument, meta, sql);
       if (definitions != null && Array.isArray(definitions)) {
         definitions.forEach((definition) => {
-          permissionContext.gqlDocument.definitions.push(definition);
+          (permissionContext.gqlDocument.definitions as DefinitionNode[]).push(definition);
         });
       }
     }
